Replace Java internal serialisation with protocol buffers.
This is useful so that the format can change over time and be compatible with existing data stored in the KeyValueDatastore container that Kafka provides. If upgrades are not possible, then there would need to be some sort of migration process, which would complicate deployments.

Also moved the KeySerde and ValuesSerde stuff out of the AccumulatorProcessor class as they were getting pretty big and cluttering it up.