Add Java datastore process.
The process reads one or more input files from the local file system or S3, processes and buckets them, and writes the resulting data to a FlatBuffers file or an ORC file, again in either S3 or the local filesystem.

The intention is that this is called one or more times for each time a file is uploaded which covers a `(tile, hour)` tuple that is wanted in the output.